// src/api/types.ts
export type DataType = {
    suggestions: ItemType[]
}

export type ItemType = {
    data: {
        area: string
        area_fias_id: string
        area_kladr_id: string
        area_type: string
        area_type_full: string
        area_with_type: string
        beltway_distance: string
        beltway_hit: string
        block: string
        block_type: string
        block_type_full: string
        capital_marker: string
        city: string
        city_area: string
        city_district: string
        city_district_fias_id: string
        city_district_kladr_id: string
        city_district_type: string
        city_district_type_full: string
        city_district_with_type: string
        city_fias_id: string
        city_kladr_id: string
        city_type: string
        city_type_full: string
        city_with_type: string
        country: string
        country_iso_code: string
        divisions: string
        entrance: string
        federal_district: string
        fias_actuality_state: string
        fias_code: string
        fias_id: string
        fias_level: string
        flat: string
        flat_area: string
        flat_cadnum: string
        flat_fias_id: string
        flat_price: string
        flat_type: string
        flat_type_full: string
        floor: string
        geo_lat: string
        geo_lon: string
        geoname_id: string
        history_values: string
        house: string
        house_cadnum: string
        house_fias_id: string
        house_kladr_id: string
        house_type: string
        house_type_full: string
        kladr_id: string
        metro: string
        okato: string
        oktmo: string
        postal_box: string
        postal_code: string
        qc: string
        qc_complete: string
        qc_geo: string
        qc_house: string
        region: string
        region_fias_id: string
        region_iso_code: string
        region_kladr_id: string
        region_type: string
        region_type_full: string
        region_with_type: string
        room: string
        room_cadnum: string
        room_fias_id: string
        room_type: string
        room_type_full: string
        settlement: string
        settlement_fias_id: string
        settlement_kladr_id: string
        settlement_type: string
        settlement_type_full: string
        settlement_with_type: string
        source: string
        square_meter_price: string
        stead: string
        stead_cadnum: string
        stead_fias_id: string
        stead_type: string
        stead_type_full: string
        street: string
        street_fias_id: string
        street_kladr_id: string
        street_type: string
        street_type_full: string
        street_with_type: string
        tax_office: string
        tax_office_legal: string
        timezone: string
        unparsed_parts: string
    }
    unrestricted_value: string
    value: string
}